+++ resources/assets/js/components/home.js
@@ -1,11 +1,13 @@
 import React from 'react';
+import LoginForm from './auth/login-form';
 
 module.exports = React.createClass({
 
 	render: () => {
 		return(
-			<div className="freelane-home">
+			<div className="freelance-home">
 				<p>This is the home view</p>
+				<LoginForm />
 			</div>
 		);
 	}