+++ resources/assets/js/main.js
@@ -1,7 +1,7 @@
 // React components, react router, etc.
 import React from 'react';
 import { render } from 'react-dom';
-import { Router, Route, IndexRoute, Link, hashHistory } from 'react-router';
+import { Router, Route, IndexRoute, Link, browserHistory } from 'react-router';
 
 // App components
 import Freelance from './freelance';
@@ -9,10 +9,10 @@ import Home from './components/home';
 import Clients from './components/clients/clients';
 
 render(
-	<Router history={hashHistory}>
+	<Router history={browserHistory}>
 		<Route path="/" component={Freelance}>
 			<IndexRoute component={Home} />
 			<Route path="clients" component={Clients} />
 		</Route>
 	</Router>
-, document.getElementById('freelance-app'));``
+, document.getElementById('freelance-app'));