+++ resources/views/app.blade.php
@@ -1,0 +1,6 @@
+@extends('layouts.app')
+
+@section('content')
+	<!-- React app component -->
+	<div id="freelance-app"></div>
+@endsection